DZ2-9/hw_scrapy/pyproject.toml
[tool.poetry]
name = "hw-scrapy"
version = "0.1.0"
description = ""
authors = ["NikYurchik <14urka@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_scrapy"}]

[tool.poetry.dependencies]
python = "^3.10"
scrapy = "^2.10.0"
mongoengine = "^0.27.0"
python-dateutil = "^2.8.2"
redis-lru = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
